+++ CHANGELOG.md
@@ -19,8 +19,11 @@ v 2019.4 (?) Changes in this release:
 - Added logging to CRUD methods of handler and a diff method with context
 - HTTP errors are logged at DEBUG level only (#1282)
 - Verify hashes when serving a file (#532)
+- Mark resource as failed when code loading fails (#1520)
 - Print extra env variables in init log and only store those in database (#1482)
 - Add feature manager for enabling and disabling orchestrator features (#1530)
+- Add get_environment_id to plugin context (#1331)
+- Log server bind address and bind port on startup
 
 IMPORTANT CHANGES:
 - Older compiler versions are no longer supported with this server